requirements.txt
Flask==3.0.3
huggingface-hub==0.23.0
python-dotenv==1.0.1
